view로직 분리
