+++ src/main/java/Car.java
@@ -9,14 +9,6 @@ public class Car {
     static final int CAR_NAME_LENGTH_BOUND = 5;
     static final int CAR_MOVE = 4;
 
-    Scanner scanner = new Scanner(System.in);
-
-    public void enterCarsName() {
-        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-        String carsName = scanner.next();
-        splitCarsName(carsName);
-    }
-
     public void splitCarsName(String carsName) {
         String[] carsList = carsName.split(",");
         for (int i = 0; i < carsList.length; i++) {
